Add a prelude module to remove common boilerplate from projects
This should simplify the process of kicking off a project with nannou.

At the moment I've just picked a few items which are already showing up
often in the examples. I don't have a particular rule in mind yet for
the kinds of items that should/shouldn't be added to the prelude. Maybe
we can come up with a set of rules for this, but for now I'm sure we
should be able to pick these pretty intuitively.

I guess this can be thought of as a more conservative `#include
ofMain.h` or Processing's list auto-included items and keywords.

This closes #10.