Remove `Other` variant from `SimpleWindowEvent` in favour of `Option`.
I think this makes a bit more sense semantically, rather than just
delivering a meaningless `Other` variant for every raw event that cannot
be interpreted as a `SimpleWindowEvent`.

This also improves the `simple_window` example where we print out each
emitted `SimpleWindowEvent`, as we no longer have `Other` spammed
between all the events that actually mean something.

There is a slight trade-off here in that the pattern matching is not
*quite* as simple as it was originally, however I think what's actually
going on should hopefully still be pretty obvious?

@freesig @JoshuaBatty