Add support for audio output streams!
Allows for:

- Enumerating available devices
- Creating a new audio output stream with custom device, sample rate,
channels, sample format and buffer size.

The API works by creating a stream with some given state that is to live
on the audio thread along with a function that renders the state to an
audio buffer, returning the result of both. The user can also send
updates to the audio state from the GUI thread via the stream handle
using the `send` method.

A `simple_audio.rs` example has been added which renders a sine wave to
the default output stream. @freesig / @joshuabatty any chance either of
you chaps could test the example on macOS? It works on linux fine, just
(except `pause` and `unpause` haven't been implemented properly in
cpal), curious to hear if cpal is working sweet on macos atm.