Add **Draw** API - simplified graphics API
Currently only supports ellipses, but the rest of the geometry types
should be easy to add.

Also provides a background method for a simpler way of clearing the gl
frame.

Provides methods for yielding raw vertices, indices and transformed
vertices.

Still need to add support for text, images and other types of textures.