Add simplified App constructors. Support a default App window.
This should simplify the process of working on a single window App.
Users can call `nannou::Builder::run_window` to run the app with an
added default window. The `nannou::run` and `nannou::view` functions now
do this internally.

`nannou::view` was added to make it easier to get started with
graphics-only apps. Much friendlier to the creative process and for
quick experiments.