Add simple accessor field for getting the time since app start in secs
From the docs:

The time in seconds since the `App` started running.

Primarily, this field is a convenience that removes the need to call
`app.duration.since_start.secs()`. Normally we would try to avoid using
such an ambiguous field name, however due to the sheer amount of use
that this value has we feel it is beneficial to provide easier access.

This value is of the same type as the scalar value used for describing
space in animations.  This makes it very easy to animate graphics and
create changes over time without having to cast values or repeatedly
calculate it from a `Duration` type. A small example might be
`app.time.sin()` for simple oscillation behaviour.

**Note:** This is suitable for use in short sketches, however should be
avoided in long running installations. This is because the "resolution"
of `f64` values reduces the higher the number becomes. Instead, we
recommend using `app.duration.since_start` or `app.duration.since_last`
to access a more precise form of this time.