Improve the efficiency of the App proxy by only calling OS when needed
The app proxy now checks a flag to determine whether or not to call the
`winit::EventsLoopProxy::wakeup` method and in turn the underlying OS
wakeup function. This flag indicates whether or not the event loop is
currently "asleep" aka waiting for input events. The proxy will now only
wakeup the main event loop if the flag indicates that the event loop is
currently asleep.

This should allow the user to call the `wakeup` method as often as they
want without unnecessarily triggering the underlying OS wakeup function
which on some systems can cause lock contention and significant CPU
increase.