Updated the simple_ui example
changed from using conrod widget to draw a background colour to using
nannou to draw an ellipse. There are sliders for controlling scale,
rotation and the resolution of the shape. A button triggers a random
colour for the shape and a XYpad allows the use to position the shape on
the canvas.