Add `draw.mesh()` API
Allows for drawing a mesh via a sequence of triangles. The triangles can
be provided either as a sequence of `Tri<V>` or via a sequence of
vertices alongside a sequence of indices into those vertices.

The `IntoDrawn` trait API has been changed to allow for a more flexible
`Indices` type. The `Indices` type now has access to the Draw's
intermediary mesh each time the next index is requested.