Update `Wait` and `Rate` loop modes for switch to Vulkan
This involved refactoring the `RefreshSync` run loop into a set of
modular functions for easier re-use between loop modes.

Note that while each loop mode is now supported, the `Wait` and `Rate`
present modes are not yet optimal. If using the `Wait` or `Rate` we
should attempt to use the `Mailbox` present mode with triple-buffering
in order to avoid blocking when acquiring an image. This is different to
the `RefreshSync` mode as the `RefreshSync` mode is *driven* by the rate
at which swapchain images can be retrieved for each window.