Allow Swapchain customisation. Recreate Swapchain on LoopMode change.
This allows for specifying a set of custom swapchain parameters during
the window building process.

The swapchain will now be recreated each time the `LoopMode` is changed
at runtime. It will attempt to use optimal `PresentMode` and image count
for the new swapchain unless the user has chosen a specific value for
either of these parameters.

The `loop_mode` example has been updated with some short docs along with
the addition of the new loop mode. The window title now displays the
current `LoopMode`.