Add `vk_teapot.rs` example
This should act as a nice simple example for getting started with 3D,
and will also be a useful reference for testing the 3D component of the
`draw` API.

I'm thinking it might be nice to follow this up with a
`vk_teapot_camera.rs` example which is basically the exact same example
but allows you to move around a virtual camera using WASD and orient it
using the mouse.

Another example I'd like to make is a `vk_teapot_matrices.rs` example
which is the exact same but with a GUI overlay that allows you to tweak
the world, view and projection matrices directly in order to gain an
intuition for how they work.