Create a `SwapchainFramebuffers` type
This significantly simplifies the process of creating and maintaining
swapchain image framebuffers, currently probably the most tedious part
of using raw vulkan with nannou. See the docs for the new
`window::SwapchainFramebuffers` type and its `update` method for more
details on how it works under the hood.

This significantly simplifies the `ui` and `draw` backends, as well as
the raw vulkan examples as a result. The easiest way to see the impact
of this PR is probably to check the diff for the vk examples.

cc @JoshuaBatty should be a good start on this simplification process!