Recreate window swapchains if user model requires loop change
Previously this check was missing, so if a user changed the loop mode
after they created a window, the swapchain would not always match
accordingly until the loop mode changed again. Now, we check for a
change to the loop mode after the user model function is called and
recreate window swapchains if necessary.