Remove `state::time::Duration` in favour of a `time::DurationF64` trait.
The idea behind the `Duration` type was to extend the std `Duration`
with simple methods for retrieving useful time units with a `f64`
representation. This approach was unwieldy and confusing to use in
practise as it required constantly converting between the `std` and
nannou `Duration` types.

Having the new `DurationF64` trait in the prelude should simplify things
by extending the `std::time::Duration` type the methods directly.