Fallback to `Fifo` present mode if `Mailbox` is unavailable
Nannou will now fall back to the `Fifo` present mode if `Mailbox` is
not supported by the swapchain when switching the `Wait` or `Rate`
`LoopMode`s.

This should fix the issue on macos where certain examples that use the
`Wait` or `Rate` loop modes would `panic!` with an
`UnsupportedPresentMode` error.

Closes #262.

@freesig could you confirm that this fixes the `simple_ui.rs` example
for you? It's probably also worth checking that the `loop_mode.rs` works
now too as that example switches between all of the different loop modes
one at a time on keypress.