Embrace the `vk` prefix. Re-export vulkano via a new `vk` module.
Also removes the `gpu` module in favour of the new `vk` module.

Many vulkano items have been re-exported into the `vk` module as well to
avoid name repetition and ease access to common types. Items that would
have an ambiguous meaning if they were to be re-exported have been left
out of the re-export list.

See the examples for how to update!