Add an `examples/laser` directory demonstrating `lasy` laser streams.
Recently I started a new crate for discovering and streaming to laser
DACs called [**lasy**][1]. This commit adds two examples demonstrating
how to work with both the RawStream and FrameStream APIs from the new
crate.

Originally I started integrating this directly into nannou itself, but
was weary of adding the extra compile-time work when the vast majority
of users probably don't have access to lasers. We can reconsider if we
can find a good motivation to do so, but otherwise I think the new
examples already demonstrate a decent level of simplicity!

[1]: https://github.com/nannou-org/lasy