Edits to the Anatomy of a Nannou App chapter
This started as a series of small edits to update @tpltnt's chapter from
rust 2015 to 2018 and from nannou 0.8 to 0.9, but turned into quite a
large edit.

- Updates from rust 2015 -> 2018.
- Updates from nannou 0.8 -> 0.9.
- Added a "Tutorial Info" section to the top with "Author", "Required
Knowledge" and "Reading Time" (might be nice to make this consistent
across all tutorials).
- Formats code for automated testing by rust-skeptic and travis.
- Rather than referring to code line numbers, we now inline each section
of code into the relevant section of the tutorial.
- A few spelling mistakes have been fixed and some chapters have been
reworded.
- The tutorial has been broken into sub-headings where each individual
snippet of code is discussed in more detail.
- Extra details have been separated into separate `Note` sections at the
end of each sub-heading.
- Added concluding remarks section.

I've also moved this chapter into a `tutorials/basics` directory and
linked it as the first tutorial under the basics section in
`tutorials.md`.