Only draw on updates. Update app timer before events. Update docs.
This makes a few small tweaks to nannou-org/nannou#353

- Make sure that we don't re-draw on every window/app/input event. That
  is, only draw if we have emitted an update.
- Move the app timer to update before emitting events so that the
  `app.time` field is up-to-date when handling those events. This should
  probably be fixed in the `Wait` loop mode as well, but that can wait
  for a future PR.
- Add documentation for the new `NTimes` loop mode.
- Run `cargo fmt` to keep travis happy.

For some reason I'm getting a `panic!` in the `loop_mode.rs` example
soon after switching from the `NTimes` loop mode to the `Wait` loop
mode, though I'm not sure if it's related to this PR. I'll do some
investigating.

In the meantime, feel free to merge this if you're happy with the tweaks
and your PR should update automatically.