WIP - Update CPAL, API -> Host, Move render/capture to builder methods.
This commit is in anticipation of the iminent CPAL update which adds a
`Host` API allowing users to select between multiple available hosts on
the platform. Namely, a feature has been added to allow support for ASIO
on Windows.

The `Api` type has been renamed to `Host`, following suit with the new
`Host` abstraction in CPAL.

The `render` and `capture` arguments to the respective
`build_output_stream` and `build_input_stream` methods have been moved
to builder methods on `input` and `output` builder types. This has been
done in order to allow the user the option to either submit a function
for processing the audio directly (same as the current behaviour) or
submitting a callback for handling stream results as allowed by the new
CPAL API.

Still waiting on tomaka/cpal#294 to be solved before merging this with
master, but at least for now I can start testing this branch with
downstream projects.