Add `draw.path()` API
This adds a path drawing API that is a vastly more flexible and correct
version of the existing polyline and polygon APIs.

Allows for tessellating any given path with either lyon's
FillTessellator or the StrokeTessellator, with builder methods provided
for all the available options.

Colored vertices are currently only supported for stroke tessellation as
I'm currently unsure how to retain the color attributes for vertices
during the fill tessellation process (see nical/lyon#488).