Make stroke and fill opt builders generic. Update ellipse for lyon.
The stroke and fill builder methods are now genericly available for any
type containing a respective `StrokeOptions` or `FillOptions` field.

The `draw.ellipse()` API has been updated to take advantage of this,
providing a new `draw.stroke(color)` method along with a suite of stroke
tessellation options.

Some methods have been added to `Theme` in order to make it easier to
retrieve colors for draw primitives.