Simplify ellipse, quad, rect and tri with polygon implementation
This also adds support for stroke tessellation to each of these
primitives.

Unfortunately, this has also resulted in `quad` and `tri` losing support
for 3-dimensional points. However, this is likely better addressed in
the future when we come up with a proper 3D tessellation pipeline
similar to lyon and some sort of ambient lighting to help visualise 3D
geometry.