Use the `notosans` crate to guarantee default font availability.
This addresses the #380.

The `notosans` crate is behind a `notosans` feature which is enabled by
default. To disable this dependency (which may help to remove roughly
~800kb of total binary size), use the `--no-default-features` flag. In
the case that the `notosans` feature is disabled, the `Ui` will attempt
to load the first font discovered in `assets/fonts`. If no fonts exist,
an error will be returned.

Closes #380.