Add bool to assess validity of app.mouse.x, .y
When at least two windows are used, the first click on the unfocused window will
set the pressed state of the mouse using the position from app.mouse and also
set the app.mouse.window to the current window.
BUT in this case the position is invalid, because (I guess it is a winit
specificity) no CursorMoved event has been generated, so the position is the
last position of the cursor in the previous window. We use
`app.mouse.is_position_valid` to track the validity of position in app.mouse.