Implement `draw.text("foo")` API
This finishes the implementation of the `draw.text("foo")` API.

For now, the API is implemented in terms of the `draw.path()` API and as
a result large chunks of text are very slow to render. This will be
improved in the future by adopting a GPU glyph cache, much of which is
already implemented. This glyph cache work has been suspended in order
to come up with a nicer `Draw` render pipeline that is friendlier to the
CPU text raster process necessary in during caching.