Fix bug where draw.mesh intermediary mesh indices were incorrect
This was introduced due to a change that was made to the way
intermediary mesh indices are handled when the draw.path() API was
introduced. The behaviour should now be fixed and users should now be
able to draw any number of shapes, paths or meshes in any order without
issues.

Closes #402.