Fix an issue with fallback font detection in Ui construction
Currently, the `Ui` constructor method unnecessarily returns an Error if
the assets directory cannot be found, even if the `notosans` feature is
enabled. This commit changes the behaviour to attempt to use the
`notosans` crate font if the feature is enabled and only attempt to
discover the assets directory if the feature is disabled.