Add new `Texture` and `TextureBuilder` types.
Also re-exports the `wgpu` crate's `Texture` type as `TextureHandle`.

The new `Texture` type acts as a much more convenient handle around the
`wgpu::Texture`, providing access to the descriptor from which it was
created along with all the info it provides.

The new `TextureBuilder` type wraps the `TextureDescriptor` with a set
of reasonable defaults, allowing for more convenient texture
construction.