All for submission of `Frame` prior to the end of `view`
This allows the user to ensure that the command buffer responsible for
drawing to the swap chain texture is submitted prior to the end of their
`view` function. This is necessary to allow the user to read back
buffers and textures involved in drawing to the texture at the end of
the `view` function. Without this change, it is not possible to do so in
nannou's current API.

The most significant user-impacting change is that the `view` functions
now take the `Frame` by value, rather than by reference.