Add manual msaa resolve to TextureReshaper. Solves validation error.
Previously a validation error was occurring due to passing a sampled
texture to the bind group without declaring that the texture was
multisampled in the fragment shader.

This adds some new multisampled shaders that correctly declare the
texture and do the multisampling resolve manually.

Dedicated shaders have been added for common sample counts with unrolled
loops to avoid having the shader deal with conditions based on uniform
data. A fallback shader is provided for all other sample counts.

Closes #462.