Add `.size(w, h)` and `.run()` builder methods to `nannou::sketch`
This allows for specifying the window size of a sketch without having to
set the size within the `view` function every frame. Closes #461.

Note that as a result of `nannou::sketch(view)` producing a builder
type, you now must call `.run()` to make the sketch do anything.

The `app::Builder` now also supports a `.size(w, h)` builder method for
specifying the default window width and height in points.