Relax trait bounds on mesh types. Remove intermediary mesh types.
Originally the idea was that the intermediary mesh may store attribute
vectors with different lengths based on what data was submitted by the
user. This was never used in the end, and there is very little cost to
allocating large slices of zeroed data in the unused attributes anyway.
This greatly simplifies a lot of the `draw.mesh()` API implementation.

As a part of this, I noticed that a few of the mesh traits had
unnecessarily restrictive trait bounds that propagated up through usage
in an annoying manner. These bounds have been relaxed in order to
improve usability.

The old glyph cache has been removed now that the `draw::renderer`
declares its own type.