Add `Draw` to prelude. Update examples for removal of `app::Draw`.
Now that there is no longer any ambiguity between `draw::Draw` and
`app::Draw` since `app::Draw` was removed, we can safely include the
`draw::Draw` within the prelude. As demonstrated in the updated
examples, this type is a common addition and along with the new group
transformation API it makes sense to provide easy access to it.