Add texturing support to `draw.mesh()`, `path()` and `polygon()`
Also adds `textured_mesh.rs`, `textured_path.rs` and
`textured_polygon.rs` examples to demonstrate!

This also adds some consistency to the way that the mesh, path and
polygon APIs receive points.

The `draw.mesh()` methods:
- `points(points)` - iterator of `point`s
- `points_colored(points)` - iterator of `(point, color)`
- `points_textured(tex_view, points)` - iterator of `(point, coords)`
- `tris(tris)` - tri iterator of `point`s
- `tris_colored(tris)` - tri iterator of `(point, color)`
- `tris_textured(tex_view, tris)` - tri iterator of `(point, coords)`
- `indexed(points, indices)`
- `indexed_colored(points, indices)`
- `indexed_textured(tex_view, points, indices)`

The `draw.path()` methods:
- `points(points)` - iterator of `point`s
- `points_colored(points)` - iterator of `(point, color)`
- `points_textured(tex_view, points)` - iterator of `(point, coords)`

The `draw.polygon()` methods:
- `points(points)` - iterator of `point`s
- `points_colored(points)` - iterator of `(point, color)`
- `points_textured(tex_view, points)` - iterator of `(point, coords)`

The `IntoVertex` and `IntoPoint` traits have been removed in favour of
having each of these APIs yield tuples of the expected attributes. E.g.
the `points_colored` examples expect tuples of `(point, color)` where
`point` implements `Into<Point>` and `color` implements `IntoLinSrgba`.
Similarly, `points_textured` expect tuples of `(point, tex_coords)`.

**TextureView**

In order to enable this texturing API, we needed a way to allow for
uniquely identifying the `TextureView`s within the renderer so that we
only have to switch texture bind groups when absolutely necessary. In
order to achieve this, a wrapper `TextureView` type has been created and
the old wgpu `TextureView` type has been renamed to `TextureViewHandle`.
This follows the convention set with the `Texture` and `TextureHandle`
types. Similarly to `Texture`, the `TextureView` type now also provides
methods for accessing its descriptor, along with shorthand method for
accessing each of its fields (e.g. size, aspect, layer count, etc). A
user can produce a `TextureViewBuilder` via the `texture.view()` method.
The `TextureViewBuidler` allows for specfiying exactly what part of the
texture should be viewed, defaulting to the full texture in the case
that parameters are specified. The texture view can be built by calling
`.build()`. The `wgpu_teapot*` examples have been updated to create
their `depth_texture_view`s using this new technique and may serve as an
example.

A `Texture::upload_data` method has been added to simplify the process
of uploading a slice of data to a buffer and copying it to the texture.