Simplify the common case of loading a Texture from an image or path
Adds the following:

- `wgpu::Texture::from_image(&app, &img)` and
- `wgpu::Texture::from_path(&app, path)`.

The first argument may be any type that can provide access to a
`wgpu::Device` and `wgpu::Queue`. For general use, the `app` is probably
the easiest way of accessing these. See the constructor docs for more
information.

The `textured_*.rs` and `wgpu_image.rs` examples have been updated.