Add a `draw.texture(texture)` API.
Allows for drawing a texture to a rectangular area! By default the area
is the size of the texture.

Also allows for specifying the area that should be drawn, e.g.
`draw.texture(texture).area(rect)`. The `rect` should describe the
texture coordinates of the source texture that should be displayed.