Automate publishing nannou repositories if version is changed
This adds a workflow that will automatically try to publish the `nannou`
and `nannou_*` crates every time a commit is pushed to `master`.

The idea is that we can now run `cargo run -p set_version -- X.Y.Z`,
commit the version update and push to the repo and the github action
will automatically publish all versions that have been incremented.

This workflow may need a little tweaking over the next couple of
releases as its difficult to test until it comes time to make a release.