Support for stream error handling to nannou_laser
This adds the ability to add a stream error callback function to both
the raw and frame stream types during building.

An action can be specified to take in response to the error. Options
include attempting to re-detect the DAC, reconnect to the TCP stream and
closing the stream thread.

Methods have also been added to the stream types to allow for checking
whether or not the stream has been closed, and for closing the stream
manually and checking the result.

The `ffi` module has been updated to provide access to this new
functionality.