Add new "Basics - Window Coordinates" tutorial
This adds a new tutorial that takes a dive into how window coordinates
work in nannou. The tutorial attempts to assume no prior experience with
coordinate systems, explaining how things work in nannou from the ground
up. That said, the hope is that it will still be easy for more
experienced creative coders to quickly find the information that they
need (e.g. axis orientations, window bounds, etc) towards the beginning
of the tutorial.

The tutorial also takes a dive into using the `Rect` type to easily
achieve common tasks like alignment, padding and relative positioning.
Only a few `Rect` methods are covered - exploration of the remaining
methods are left as an exercise for the reader.

Finally we finish with a short explanation of Points vs Pixels. Though
perhaps this could be moved into a follow-up mini-tutorial?

This commit fixes a couple of bugs discovered in the `Rect`
implementation whil writing the tute. The `Window` type has also been
added to the prelude.

Definitely excited about landing this one! How to work with nannou's
coordinates is ceratinly a FAQ, hopefully we can point to this to do the
trick.

cc @madskjeldgaard, @freesig, @joshuabatty