Add a `draw.arrow()` API. Include a new `draw_arrow.rs` example
The new API is almost identical to `draw.line()`, except that you can
also specify the `.head_length(l)` and `.head_width(w)` if you wish. By
default, these are equal to `weight * 4.0` and `weight * 2.0`
respectively.

This might be particularly useful for things like vector field
visualisations or for debugging.