`nannou_isf` progress, Update isf_demo.rs with basic GUI
Thanks to @KeyboardSounds for mentioning the state of the nannou_isf
crate in #682 - I was reminded that I have a branch lying around which
made a bit more progress!

The `isf_demo.rs` example now provides a small GUI window with a list of
shader paths read from the `assets/isf` directory. You can select
between them to load an ISF pipeline for that shader. Once selected, the
GUI should show whether the ISF descriptor, fragment and vertex shaders
were parsed successfully. If everything was parsed successfully, a new
ISF pipeline is loaded for the selected ISF shader. From this point, you
should be able to edit the selected shader and have it automatically
hotload the result when you save the file from your editor.

Currently, the `assets/isf` directory contains a bunch of test shaders
provided by the vidvox crew along with the original implementation. I'm
not sure exactly what each should look like yet. Most fail to load
still, though I don't quite remember how much of this is due to certain
parts still being unimplemented in `nannou_isf`, and which parts are
bugs in wgpu 0.5 (more likely the former!). There's a lot that goes on
in the ISF pipeline, and little testing has been done so far, so I'm
glad it's working as well as it does currently. I won't get a chance to
dive much deeper for another month or two, but if you're interested in
contributing to `nannou_isf`, doing some investigating into this could
be a really nice way to do so while learning how wgpu works! It may be
worth landing #665 first as wgpu 0.6 has quite a bit more validation
and error checking going on which could be useful to debug crashes that
otherwise might be quite confusing in 0.5.