Fix empty mesh panic
Previously, if an empty Draw is passed to Renderer::encode_render_pass, it would attempt to create empty buffers. This panics because wgpu::Device::create_buffer_mapped asserts size != 0. This fix prevents a panic in some examples, though it could probably be made cleaner if the mesh itself had an is_empty method.