Add CONTRIBUTING.md. Add Contributing chapter to guide.
This new chapter provides the following:

- A small overview of nannou's codebase and how it is managed (git and
  PRs).
- A **PR Checklist** page for use as a reference/reminder before opening
  PRs.
- A **PR Reviews** page for tips on how to speed up reviews and make
  sure they go smoothly.
- A **Publishing New Versions** page to inform contributors how they can
  get a new version of nannou published.

I think a lot of this could be improved and fleshed out a little
further, but hopefully this makes for a decent start!

Related to #553 but still some good ideas there that are not quite
implemented. Maybe it could be closed in favour of opening up specific
issues for each "contributing chapter" feature?