Publish version 0.15
See the changelog below for an overview of the changes included in this
version.

For the most part, this is an update for wgpu 0.5. @alekspickle
already has a WIP for wgpu 0.6 at #655! I'm more than happy to publish a
new version again as soon as that lands, but figured this one is already
well overdue :)

Also, find out how you can get new nannou versions published quicker in
the new Contributing chapter of the guide
[here](https://guide.nannou.cc/contributing/publishing-new-versions.html)).