Add `TextureViewInfo` type for `TextureView`
This adds a `TextureViewInfo` that fulfills the role for `TextureView`
that the `TextureViewDescriptor` used to.

In 0.6, the `TextureViewDescriptor` was changed so that a few of the
fields are optional. This is useful for users of the raw wgpu API, as it
allows for fields that may be trivially inferred from the parent texture
to do just that during the texture view creation process.

Nannou originally stored the `TextureViewDescriptor` in the
`TextureView` type in order to allow the user to access information
about the texture view without having to manually pipe through all the
same information that was used to construct it. The newly introduced
optional fields are inconvenient in this case, as we want to store the
known, built values.

The new `TextureViewInfo` is almost identical to the
`TextureViewDescriptor`, but differs in that all optional arguments that
may be trivially inferred during texture view creation are no longer
optional.