Publish 0.16 - Update changelog with related changes
It's time!

Notable changes include the update from wgpu 0.5 to 0.7 and a bunch of
bug fixes for graphics related stuff. See the `guide/src/changelog.md`
diff for a detailed list of the changes included in this release.

I think I mentioned on matrix that I was hoping to get #684 into this
release, but then realised that `nannou_isf` is still unpublished
anyway, so we can publish it at 0.16 if/when those changes are ready.