Fix missing texture bindings and recreate pipeline on texture changes
Fixes an issue where a texture would be omitted from the bind group
layout if it was not yet loaded. This is not OK, as the shader code
likely references the texture, meaning something has to be there whether
or not it is loaded. To work around this, we now use a placeholder
texture.

We now properly compare the full list of texture descriptors when
checking whether or not the bind group and pipeline need to be
recreated. Previously we only checked the texture count, which means
that if a placeholder texture was replaced with an actual texture the
actual texture would never be loaded.