fix: NOC 3_04 exercise to work on Windows
When running the 3_04 exercise on Windows, the background is completely
black, so the spiral path is not visible. I tested on my mac, the
background, without being set, is green, so it kinda works on macOS.
Nevertheless, this commit sets the background to white on the first
frame so it can work on both Windows and macOS and be consistent with
the Nature of Book's example, which has the white background as well.