Refactor `Rect` and `Cuboid` methods for `Point*` and `Vec*` types
This refactors the `Rect` and `Cuboid` methods to try to use `Point*`
and `Vec*` types where possible for `Rect<f32>` and `Rect<f64>` in order
to maintain an API closer to the original and avoid large sweeps amounts
of breaking changes.

Also exposes common `glam` types via the `nannou_core` `prelude` module.