Improve `safe_file_save` behaviour, address some edge cases
This fixes a couple of issues. Previously, it was possible that the
written contents were not yet fully synchronised with the temporary file
on disk before the tmp file was renamed to the new file. We now ensure
that this synchronisation is performed and return an error if it fails
for some reason.

Previously, we removed the original file before replacing it with the
temporary file. This meant that data could be lost if an error occurred
while renaming the temporary file. Now, we rename the original to
`.backup`, then we replace it with the temporary file, and only after
replacement completes successfully do we delete the `.backup` file.